Move user profile cache to its own module.
Require functions that must run on the main UI thread to accept
a reference to `Cx` as a guarantee that they are running
on the main thread, as `&mut Cx` is only available on the main UI thread.

(untested)
