+++ src/profile/mod.rs
@@ -3,6 +3,7 @@ use makepad_widgets::Cx;
 pub mod my_profile_screen;
 pub mod profile_screen;
 pub mod user_profile;
+pub mod user_profile_cache;
 
 pub fn live_design(cx: &mut Cx) {
     profile_screen::live_design(cx);
