+++ src/profile/user_profile.rs
@@ -1,153 +1,12 @@
-use std::{borrow::Cow, cell::RefCell, collections::{btree_map::Entry, BTreeMap}, ops::{Deref, DerefMut}, sync::Arc};
-use crossbeam_queue::SegQueue;
+use std::{borrow::Cow, ops::{Deref, DerefMut}, sync::Arc};
 use makepad_widgets::*;
-use matrix_sdk::{room::{RoomMember, RoomMemberRole}, ruma::{events::room::member::MembershipState, OwnedRoomId, OwnedUserId, UserId}};
+use matrix_sdk::{room::{RoomMember, RoomMemberRole}, ruma::{events::room::member::MembershipState, OwnedRoomId, OwnedUserId}};
 use crate::{
     shared::avatar::AvatarWidgetExt, sliding_sync::{get_client, is_user_ignored, submit_async_request, MatrixRequest}, utils
 };
 
-thread_local! {
-    /// A cache of each user's profile and the rooms they are a member of, indexed by user ID.
-    static USER_PROFILE_CACHE: RefCell<BTreeMap<OwnedUserId, UserProfileCacheEntry>> = RefCell::new(BTreeMap::new());
-}
-struct UserProfileCacheEntry {
-    user_profile: UserProfile,
-    room_members: BTreeMap<OwnedRoomId, RoomMember>,
-}
+use super::user_profile_cache::{self, get_user_profile_and_room_members, get_user_room_member_info};
 
-/// The queue of user profile updates waiting to be processed by the UI thread's event handler.
-static PENDING_USER_PROFILE_UPDATES: SegQueue<UserProfileUpdate> = SegQueue::new();
-
-/// Enqueues a new user profile update and signals the UI
-/// such that the new update will be handled by the user profile sliding pane widget.
-pub fn enqueue_user_profile_update(update: UserProfileUpdate) {
-    PENDING_USER_PROFILE_UPDATES.push(update);
-    SignalToUI::set_ui_signal();
-}
-
-/// A user profile update, which can include changes to a user's full profile
-/// and/or room membership info.
-pub enum UserProfileUpdate {
-    /// A fully-fetched user profile, with info about the user's membership in a given room.
-    Full {
-        new_profile: UserProfile,
-        room_id: OwnedRoomId,
-        room_member: RoomMember,
-    },
-    /// An update to the user's room membership info only, without any profile changes.
-    RoomMemberOnly {
-        room_id: OwnedRoomId,
-        room_member: RoomMember,
-    },
-    /// An update to the user's profile only, without changes to room membership info.
-    UserProfileOnly(UserProfile),
-}
-impl UserProfileUpdate {
-    /// Returns the user ID associated with this update.
-    #[allow(unused)]
-    pub fn user_id(&self) -> &UserId {
-        match self {
-            UserProfileUpdate::Full { new_profile, .. } => &new_profile.user_id,
-            UserProfileUpdate::RoomMemberOnly { room_member, .. } => &room_member.user_id(),
-            UserProfileUpdate::UserProfileOnly(profile) => &profile.user_id,
-        }
-    }
-
-    /// Applies this update to the given user profile pane info,
-    /// only if the user_id and room_id match that of the update.
-    ///
-    /// Returns `true` if the update resulted in any actual content changes.
-    #[allow(unused)]
-    fn apply_to_current_pane(&self, info: &mut UserProfilePaneInfo) -> bool {
-        match self {
-            UserProfileUpdate::Full { new_profile, room_id, room_member } => {
-                if info.user_id == new_profile.user_id {
-                    info.avatar_info.user_profile = new_profile.clone();
-                    if &info.room_id == room_id {
-                        info.room_member = Some(room_member.clone());
-                    }
-                    return true;
-                }
-            }
-            UserProfileUpdate::RoomMemberOnly { room_id, room_member } => {
-                log!("Applying RoomMemberOnly update to user profile pane info: user_id={}, room_id={}, ignored={}",
-                    room_member.user_id(), room_id, room_member.is_ignored(),
-                );
-                if info.user_id == room_member.user_id() && &info.room_id == room_id {
-                    log!("RoomMemberOnly update matches user profile pane info, updating room member.");
-                    info.room_member = Some(room_member.clone());
-                    return true;
-                }
-            }
-            UserProfileUpdate::UserProfileOnly(new_profile) => {
-                if info.user_id == new_profile.user_id {
-                    info.avatar_info.user_profile = new_profile.clone();
-                    return true;
-                }
-            }
-        }
-        false
-    }
-
-    /// Applies this update to the given user profile info cache.
-    fn apply_to_cache(self, cache: &mut BTreeMap<OwnedUserId, UserProfileCacheEntry>) {
-        match self {
-            UserProfileUpdate::Full { new_profile, room_id, room_member } => {
-                match cache.entry(new_profile.user_id.to_owned()) {
-                    Entry::Occupied(mut entry) => {
-                        let entry_mut = entry.get_mut();
-                        entry_mut.user_profile = new_profile;
-                        entry_mut.room_members.insert(room_id, room_member);
-                    }
-                    Entry::Vacant(entry) => {
-                        entry.insert(UserProfileCacheEntry {
-                            user_profile: new_profile,
-                            room_members: {
-                                let mut room_members_map = BTreeMap::new();
-                                room_members_map.insert(room_id, room_member);
-                                room_members_map
-                            },
-                        });
-                    }
-                }
-            }
-            UserProfileUpdate::RoomMemberOnly { room_id, room_member } => {
-                match cache.entry(room_member.user_id().to_owned()) {
-                    Entry::Occupied(mut entry) =>{
-                        entry.get_mut().room_members.insert(room_id, room_member);
-                    }
-                    Entry::Vacant(entry) => {
-                        // This shouldn't happen, but we can still technically handle it correctly.
-                        warning!("BUG: User profile cache entry not found for user {} when handling RoomMemberOnly update", room_member.user_id());
-                        entry.insert(UserProfileCacheEntry {
-                            user_profile: UserProfile {
-                                user_id: room_member.user_id().to_owned(),
-                                username: None,
-                                avatar_img_data: None,
-                            },
-                            room_members: {
-                                let mut room_members_map = BTreeMap::new();
-                                room_members_map.insert(room_id, room_member);
-                                room_members_map
-                            },
-                        });
-                    }
-                }
-            }
-            UserProfileUpdate::UserProfileOnly(new_profile) => {
-                match cache.entry(new_profile.user_id.to_owned()) {
-                    Entry::Occupied(mut entry) => entry.get_mut().user_profile = new_profile,
-                    Entry::Vacant(entry) => {
-                        entry.insert(UserProfileCacheEntry {
-                            user_profile: new_profile,
-                            room_members: BTreeMap::new(),
-                        });
-                    }
-                }
-            }
-        }
-    }
-}
 
 /// Information retrieved about a user: their displayable name, ID, and avatar image.
 #[derive(Clone, Debug)]
@@ -179,19 +38,19 @@ impl UserProfile {
 }
 
 
-/// Information needed to display an avatar widget: a user profile and room ID.
+/// Basic info needed to populate the contents of an avatar widget.
 #[derive(Clone, Debug)]
-pub struct AvatarInfo {
+pub struct UserProfileAndRoomId {
     pub user_profile: UserProfile,
     pub room_id: OwnedRoomId,
 }
-impl Deref for AvatarInfo {
+impl Deref for UserProfileAndRoomId {
     type Target = UserProfile;
     fn deref(&self) -> &Self::Target {
         &self.user_profile
     }
 }
-impl DerefMut for AvatarInfo {
+impl DerefMut for UserProfileAndRoomId {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.user_profile
     }
@@ -539,19 +398,19 @@ live_design! {
 
 #[derive(Clone, DefaultNone, Debug)]
 pub enum ShowUserProfileAction {
-    ShowUserProfile(AvatarInfo),
+    ShowUserProfile(UserProfileAndRoomId),
     None,
 }
 
 /// Information needed to populate/display the user profile sliding pane.
 #[derive(Clone, Debug)]
 pub struct UserProfilePaneInfo {
-    pub avatar_info: AvatarInfo,
+    pub avatar_info: UserProfileAndRoomId,
     pub room_name: String,
     pub room_member: Option<RoomMember>,
 }
 impl Deref for UserProfilePaneInfo {
-    type Target = AvatarInfo;
+    type Target = UserProfileAndRoomId;
     fn deref(&self) -> &Self::Target {
         &self.avatar_info
     }
@@ -627,29 +486,27 @@ impl Widget for UserProfileSlidingPane {
             return;
         }
 
-        // Handle the user profile info being updated by a background task.
-        let mut redraw_this_pane = false;
+        // A UI Signal indicates this user profile was updated by a background task.
         if let Event::Signal = event {
-            USER_PROFILE_CACHE.with_borrow_mut(|cache| {
-                while let Some(update) = PENDING_USER_PROFILE_UPDATES.pop() {
-                    // Insert the updated info into the cache
-                    update.apply_to_cache(cache);
-                }
+            user_profile_cache::process_user_profile_updates(cx);
 
-                // Apply this update to the current user profile pane (if relevant).
-                if let Some(our_info) = self.info.as_mut() {
-                    if let Some(new_info) = cache.get(&our_info.user_id) {
-                        our_info.user_profile = new_info.user_profile.clone();
-                        our_info.room_member = new_info.room_members.get(&our_info.room_id).cloned();
-                        redraw_this_pane = true;
-                    }
+            // Re-fetch the currently-displayed user profile info from the cache in case it was updated.
+            let mut redraw_this_pane = false;
+            if let Some(our_info) = self.info.as_mut() {
+                if let (Some(new_profile), room_member) = get_user_profile_and_room_members(
+                    cx,
+                    &our_info.user_id,
+                    &our_info.room_id,
+                ) {
+                    our_info.user_profile = new_profile;
+                    our_info.room_member = room_member;
+                    redraw_this_pane = true;
                 }
-            });
-        }
-
-        if redraw_this_pane {
-            // log!("Redrawing user profile pane due to user profile update");
-            self.redraw(cx);
+            }
+            if redraw_this_pane {
+                // log!("Redrawing user profile pane due to user profile update");
+                self.redraw(cx);
+            }
         }
 
         let Some(info) = self.info.as_ref() else { return };
@@ -754,16 +611,16 @@ impl UserProfileSlidingPane {
     /// retrieve the user's room membership info from the local cache.
     /// If the user's room membership info is not found in the cache,
     /// it will submit a request to asynchronously fetch it from the server.
-    pub fn set_info(&mut self, mut info: UserProfilePaneInfo) {
+    pub fn set_info(&mut self, _cx: &mut Cx, mut info: UserProfilePaneInfo) {
         if info.room_member.is_none() {
-            USER_PROFILE_CACHE.with_borrow(|cache| {
-                if let Some(entry) = cache.get(&info.user_id) {
-                    if let Some(room_member) = entry.room_members.get(&info.room_id) {
-                        log!("Found user {} room member info in cache", info.user_id);
-                        info.room_member = Some(room_member.clone());
-                    }
-                }
-            });
+            if let Some(room_member) = get_user_room_member_info(
+                _cx,
+                &info.user_id,
+                &info.room_id,
+            ) {
+                log!("Found user {} room member info in cache", info.user_id);
+                info.room_member = Some(room_member.clone());
+            }
         }
         if info.room_member.is_none() {
             log!("Did not find user {} room member info in cache, fetching from server.", info.user_id);
@@ -793,9 +650,9 @@ impl UserProfileSlidingPaneRef {
     }
 
     /// See [`UserProfileSlidingPane::set_info()`]
-    pub fn set_info(&self, info: UserProfilePaneInfo) {
+    pub fn set_info(&self, _cx: &mut Cx, info: UserProfilePaneInfo) {
         let Some(mut inner) = self.borrow_mut() else { return };
-        inner.set_info(info);
+        inner.set_info(_cx, info);
     }
 
     /// See [`UserProfileSlidingPane::show()`]
